Add app source and build configuration
- Add base app configuration;
The web application will feature 2 API endpoints: /add_satellite and /get_satellites.
When an external device hits /add_satellite with the required JSON schema,
a satellite record will be created and inserted into the RDS PostgreSQL
table defined in config.py. The S3 bucket (provisioned by terraform) and the
file containing the telemetry data will also be updated with the new record (the new satellite).
On the other hand /get_satellites will just be used to extract the data from the database,
in case someone decides to check on what records are inside the database. I also plan
to use the former endpoint for the lambda and further telemetry processing.

- Add app build configuration;
In order to deploy the application properly in an AWS AppRunner service,
we need to have a container (Docker) image pushed into an ECR repository.
For now, I've added the base Dockerfile that can be used to create an application image.
This will also be included in the TF code later.